Feature ETP-489: Modularize the component
